Multi-agent CL/agents/research_agent.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict
import json

class ResearchAgent:
    def __init__(self):
        self.sources = {
            'linkedin': 'https://www.linkedin.com/company/',
            'crunchbase': 'https://www.crunchbase.com/organization/',
            'glassdoor': 'https://www.glassdoor.com/Overview/'
        }

    def gather_company_data(self, industry_domains: List[str], countries: List[str]) -> List[Dict]:
        """
        Gather company data from various sources based on industry and country
        """
        companies = []
        
        for industry in industry_domains:
            for country in countries:
                # Simulate API calls to various data sources
                # In a real implementation, you would use actual APIs
                company_data = self._fetch_company_data(industry, country)
                companies.extend(company_data)
        
        return companies

    def _fetch_company_data(self, industry: str, country: str) -> List[Dict]:
        """
        Fetch company data from various sources
        """
        # This is a placeholder implementation
        # In a real implementation, you would:
        # 1. Make actual API calls to LinkedIn, Crunchbase, etc.
        # 2. Parse and validate the data
        # 3. Handle rate limiting and errors
        
        sample_data = [
            {
                'name': f'Sample Company {industry}',
                'industry': industry,
                'country': country,
                'size': '100-500',
                'website': 'https://example.com',
                'description': f'A {industry} company based in {country}',
                'source': 'linkedin'
            }
        ]
        
        return sample_data

    def _parse_company_page(self, html_content: str) -> Dict:
        """
        Parse company information from HTML content
        """
        soup = BeautifulSoup(html_content, 'html.parser')
        # Implement actual parsing logic here
        return {} 